refactor pillow extensions
